sumamrizer/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sumamrizer"
version = "0.1.0"
description = "URLやテキスト、動画から内容を文字起こしして要約するツール"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "langchain",
    "openai",
    "transformers",
    "openai-whisper",
    "torch",
    "faster-whisper",  # GPU対応Whisper代替
    "requests",
    "beautifulsoup4",  # URL処理用
    "python-dotenv",   # 環境変数管理
]

[project.scripts]
sumamrizer = "sumamrizer.main:main"
